Added a Parsl translator test
