--- tests/translators/parsl/test_parsl_translator.py
@@ -1,111 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-#
-# Copyright (c) 2025 The WfCommons Team.
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-import pathlib
-import pytest
-import shutil
-import docker
-import io
-import tarfile
-import os
-import sys
-import json
-
-from wfcommons import BlastRecipe
-from wfcommons.wfbench import WorkflowBenchmark, DaskTranslator
-
-class TestParslTranslator:
-
-    def make_tarfile_of_wfcommons(self):
-        source_dir = os.getcwd() # This assumes the testing is run from the root
-        tar_stream = io.BytesIO()
-        with tarfile.open(fileobj=tar_stream, mode='w') as tar:
-            tar.add(source_dir, arcname=os.path.basename(source_dir))
-        tar_stream.seek(0)
-        return tar_stream
-
-
-    @pytest.mark.unit
-    def test_parsl_translator(self) -> None:
-
-        # Create a workflow benchmark object to generate specifications based on a recipe (in /tmp/, whatever)
-        desired_num_tasks = 45
-        benchmark_full_path = "/tmp/blast-benchmark-{desired_num_tasks}.json"
-        shutil.rmtree(benchmark_full_path, ignore_errors=True)
-        benchmark = WorkflowBenchmark(recipe=BlastRecipe, num_tasks=desired_num_tasks)
-        benchmark.create_benchmark(pathlib.Path("/tmp/"), cpu_work=100, data=10, percent_cpu=0.6)
-        with open(f"/tmp/blast-benchmark-{desired_num_tasks}.json", "r") as f:
-            generated_json = json.load(f)
-            num_tasks = len(generated_json["workflow"]["specification"]["tasks"])
-
-        # Make sure the target dir does not already exist
-        str_dirpath = "/tmp/dask_translated_workflow/"
-        dirpath = pathlib.Path(str_dirpath)
-        if dirpath.exists():
-            shutil.rmtree(dirpath)
-
-        # Perform the translation
-        sys.stderr.write("Translating workflow...\n")
-        translator = DaskTranslator(benchmark.workflow)
-        translator.translate(output_folder=dirpath)
-
-        # Pulling the Docker image
-        client = docker.from_env()
-        sys.stderr.write("Pulling Docker image...\n")
-        image_name = "wfcommons/wfcommons-testing-dask"
-        client.images.pull(image_name)
-
-        # Launch the docker container to actually run the translated workflow
-        sys.stderr.write("Starting Docker container...\n")
-        container = client.containers.run(
-            image_name,
-            "sleep infinity",
-            volumes={str_dirpath: {'bind': str_dirpath, 'mode': 'rw'}},
-            working_dir=str_dirpath,
-            tty=True,
-            detach=True
-        )
-
-        sys.stderr.write("Installing WfCommons on the container...\n")
-        # Copy the WfCommons code to it (removing stuff that should be removed)
-        target_path = '/tmp/'  # inside container
-        tar_data = self.make_tarfile_of_wfcommons()
-        container.put_archive(target_path, tar_data)
-        exit_code, output = container.exec_run("sudo /bin/rm -rf /tmp/WfCommons/build/", stdout=True, stderr=True)
-        exit_code, output = container.exec_run("sudo /bin/rm -rf /tmp/WfCommons/*.egg-info/", stdout=True, stderr=True)
-        exit_code, output = container.exec_run("sudo /bin/rm -rf /tmp/WfCommons/bin/cpu-benchmark.o", stdout=True, stderr=True)
-        exit_code, output = container.exec_run("sudo /bin/rm -rf /tmp/WfCommons/bin/cpu-benchmark", stdout=True, stderr=True)
-
-        # Install WfCommons on the container (to install wfbench and cpu-benchmark really)
-        exit_code, output = container.exec_run("sudo python3 -m pip install -e . --break-system-packages", workdir="/tmp/WfCommons", stdout=True, stderr=True)
-
-        # Copy over the wfbench and cpu-benchmark executables
-        exit_code, output = container.exec_run("sudo cp -f /tmp/WfCommons/bin/wfbench " + str_dirpath + "bin/", stdout=True, stderr=True)
-        exit_code, output = container.exec_run("sudo cp -f /tmp/WfCommons/bin/cpu-benchmark " + str_dirpath + "bin/", stdout=True, stderr=True)
-
-        # Run the dask workflow!
-        sys.stderr.write("Running the Dask workflow on the container...\n")
-        exit_code, output = container.exec_run("python ./dask_workflow.py", stdout=True, stderr=True)
-        num_completed_tasks = output.decode().count("completed!")  # TODO: This is pretty lame
-
-        # Kill the container
-        container.remove(force=True)
-
-        # Do sanity checks
-        sys.stderr.write("Checking sanity...\n")
-        assert(exit_code == 0)
-        assert(num_completed_tasks == num_tasks)
-        # TODO: Look at the (I think) generated run.json file on the container
-
-
-
-
-
-
